backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langgraph
langchain-google-genai
sqlalchemy
psycopg2-binary
